improvement: move svm predict classifier to own module
