+++ opentutor_classifier/svm/word2vec.py
@@ -4,45 +4,22 @@
 #
 # The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 #
-from dataclasses import dataclass, asdict
-from typing import Dict, List
-import yaml
+from gensim.models import KeyedVectors
+from gensim.models.keyedvectors import Word2VecKeyedVectors
+from os import path
+from typing import Dict
 
-from sklearn import svm
-
-
-@dataclass
-class ExpectationToEvaluate:
-    expectation: int
-    classifier: svm.SVC
+WORD2VEC_MODELS: Dict[str, Word2VecKeyedVectors] = {}
 
 
-@dataclass
-class InstanceExpectationFeatures:
-    ideal: List[str]
-    good_regex: List[str]
-    bad_regex: List[str]
+def find_or_load_word2vec(file_path: str) -> Word2VecKeyedVectors:
+    abs_path = path.abspath(file_path)
+    if abs_path not in WORD2VEC_MODELS:
+        WORD2VEC_MODELS[abs_path] = KeyedVectors.load_word2vec_format(
+            abs_path, binary=True
+        )
+    return WORD2VEC_MODELS[abs_path]
 
 
-@dataclass
-class InstanceConfig:
-    question: str
-    expectation_features: List[InstanceExpectationFeatures]
-
-    def __post_init__(self):
-        self.expectation_features = [
-            x
-            if isinstance(x, InstanceExpectationFeatures)
-            else InstanceExpectationFeatures(**x)
-            for x in self.expectation_features
-        ]
-
-    def write_to(self, file_path: str):
-        with open(file_path, "w") as config_file:
-            yaml.safe_dump(asdict(self), config_file)
-
-
-@dataclass
-class InstanceModels:
-    models_by_expectation_num: Dict[int, svm.SVC]
-    config: InstanceConfig
+def load_word2vec_model(path: str) -> Word2VecKeyedVectors:
+    return KeyedVectors.load_word2vec_format(path, binary=True)
